Change scrolling algorithm
Make scrolling a bit more conservative about how it places long chunks of
content that are still short enough to fit in the screen. It now behaves
as though the screen were shorter by about as much as the height of
any existing option list. This fixes a scenario that could happen where
a chunk of content would be placed correctly, but because it was just
long enough, it would then scroll partially out of view when the
options list faded away.